refactor: step3 피드백 반영
move 함수 테스트 가능하도록 변경
decide move 함수 통합
테스트 분리
결과를 리턴하지 않는 테스트 삭제
