+++ src/main/kotlin/racingcar/RandomNumberGenerator.kt
@@ -1,0 +1,10 @@
+package racingcar
+
+import java.security.SecureRandom
+
+class RandomNumberGenerator {
+    fun getRandomNumber(): Int {
+        val secureRandom = SecureRandom()
+        return secureRandom.nextInt(10)
+    }
+}
