+++ src/test/kotlin/racingcar/CarTest.kt
@@ -7,10 +7,32 @@ class CarTest {
     @Test
     fun moveTest() {
         val car = Car(0)
-        car.move()
+        car.move(5)
         assertThat(car.moveCount).isEqualTo(1)
-        car.move()
-        car.move()
+    }
+
+    @Test
+    fun moveThreeTimesTest() {
+        val car = Car(0)
+        car.move(5)
+        car.move(6)
+        car.move(8)
         assertThat(car.moveCount).isEqualTo(3)
+    }
+
+    @Test
+    fun notMoveTest() {
+        val car = Car(0)
+        car.move(2)
+        assertThat(car.moveCount).isEqualTo(0)
+    }
+
+    @Test
+    fun notMoveThreeTimesTest() {
+        val car = Car(0)
+        car.move(2)
+        car.move(1)
+        car.move(3)
+        assertThat(car.moveCount).isEqualTo(0)
     }
 }
